implementation for sx1509 keypad

+++ esphome/components/sx1509/sx1509_keypad_sensor.h
@@ -16,7 +16,7 @@ class SX1509KeypadSensor : public sensor::Sensor, public Component {
     this->rows_ = rows;
     this->cols_ = cols;
   };
-  void set_timers(uint8_t sleep_time, uint8_t scan_time, uint8_t debounce_time) {
+  void set_timers(unsigned int sleep_time, uint8_t scan_time, uint8_t debounce_time) {
     this->sleep_time_ = sleep_time;
     this->scan_time_ = scan_time;
     this->debounce_time_ = debounce_time;
@@ -29,7 +29,7 @@ class SX1509KeypadSensor : public sensor::Sensor, public Component {
   SX1509Component *parent_;
   uint8_t rows_ = {};
   uint8_t cols_ = {};
-  uint8_t sleep_time_ = {};
+  unsigned int  sleep_time_ = {};
   uint8_t scan_time_ = {};
   uint8_t debounce_time_ = {};
 };
